API Token Web - Rename files
